backend/requirements.txt
discord.py
python-dotenv
requests
aiohttp